tests/requirements.txt
# 5G K3s KubeEdge Testbed - Test Dependencies

# Core dependencies
PyYAML>=6.0
kubernetes>=29.0.0

# CLI dependencies for interactive interface
rich>=13.0.0
simple-term-menu>=1.6.0

# Optional dependencies for enhanced functionality
# Uncomment these if you want additional features:
# requests>=2.28.0        # For HTTP requests in performance tests
# psutil>=5.9.0           # For system resource monitoring
# tabulate>=0.9.0         # For better table formatting in reports
# matplotlib>=3.5.0       # For performance graphs and charts
# plotly>=5.0.0           # For interactive performance dashboards
